Changed to default dosbox.conf
The game is too dark. Any attempt to adjust gamma through menu or console
results in a segfault. This seems to be an established bug with dosbox
and is unlikely to be fixed. Segfaulting from a fullscreen window will
not restore the desktop resolution, so do not launch in fullscreen by
default.